feat: presetGroup id값 Integer로 변경

+++ pickle-pb/src/main/java/com/example/pickle_pb/presetGroup/controller/PresetGroupController.java
@@ -37,7 +37,7 @@ public class PresetGroupController {
     }
 
     @DeleteMapping
-    public ResponseEntity<CommonResDto<?>> deletePresetGroup(@RequestParam("presetGroupId") Long presetGroupId) {
+    public ResponseEntity<CommonResDto<?>> deletePresetGroup(@RequestParam("presetGroupId") Integer presetGroupId) {
         boolean result = presetGroupService.deletePresetGroup(presetGroupId);
         return ResponseEntity.status(HttpStatus.OK).body(new CommonResDto<>(1, "그룹 삭제 성공", result));
     }
